# fix perf in exp(::Matrix) (for smallish matrices)