@stevengj 2018-09-10
#29113 is about the matrix exponential, which has nothing to do with broadcasting.  In 0.6 you need to call `expm` for this.