@Keno 2018-09-10
Do we have this as a nanosoldier benchmark and didn't notice? If not, add it?